fix ui

+++ Server/Component/Basic.cc
@@ -26,7 +26,7 @@ namespace app::component
 
     void Basic::Write(bc::BinaryCoder &coder, Type const &entity, bool isCreation)
     {
-        uint32_t state = isCreation ? 0b11 : entity.m_State;
+        uint64_t state = isCreation ? 0b11 : entity.m_State;
         coder.Write<bc::VarUint>(state);
 
         if (state & 1)
