+++ Server/Component/Drop.cc
@@ -24,7 +24,7 @@ namespace app::component
 
     void Drop::Write(bc::BinaryCoder &coder, Type const &entity, bool isCreation)
     {
-        uint32_t state = isCreation ? 0b011 : entity.m_State;
+        uint64_t state = isCreation ? 0b011 : entity.m_State;
         coder.Write<bc::VarUint>(state);
 
         if (state & 1)
