+++ Server/Component/Physical.cc
@@ -25,7 +25,7 @@ namespace app::component
 
     void Physical::Write(bc::BinaryCoder &coder, Type const &entity, bool isCreation)
     {
-        uint32_t state = isCreation ? 0b11111 : entity.m_State;
+        uint64_t state = isCreation ? 0b11111 : entity.m_State;
 
         coder.Write<bc::VarUint>(state);
